장바구니 UI, 기능 완료
